requirements.txt
torch
torchvision
numpy==1.26.4
matplotlib==3.8.0
Pillow==10.2.0